Move formspec code from game.cpp to separate file
