+++ src/client/gameui.h
@@ -13,7 +13,6 @@ using namespace irr;
 class Client;
 class EnrichedString;
 class GUIChatConsole;
-class GUIFormSpecMenu;
 struct MapDrawControl;
 struct PointedThing;
 
@@ -79,15 +78,6 @@ public:
 	void toggleHud();
 	void toggleProfiler();
 
-	GUIFormSpecMenu *&updateFormspec(const std::string &formname)
-	{
-		m_formname = formname;
-		return m_formspec;
-	}
-
-	const std::string &getFormspecName() { return m_formname; }
-	GUIFormSpecMenu *&getFormspecGUI() { return m_formspec; }
-	void deleteFormspec();
 	void clearText();
 
 private:
@@ -113,9 +103,4 @@ private:
 	gui::IGUIStaticText *m_guitext_profiler = nullptr; // Profiler text
 	u8 m_profiler_current_page = 0;
 	const u8 m_profiler_max_page = 3;
-
-	// Default: "". If other than "": Empty show_formspec packets will only
-	// close the formspec when the formname matches
-	std::string m_formname;
-	GUIFormSpecMenu *m_formspec = nullptr;
 };
